+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-08-18 03:22:00
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: db2d01ee7f7ec188c379c59dcd585377533978e1
+
+## [ForwardWidgets-2kuai] - 2025-08-18 03:22:00
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 6f299d1218857dbdd5aabb263c8f3ba48f42b8bd
+
+## [ForwardWidgets-pack1r] - 2025-08-18 03:22:00
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 74ff483217877b00467f65edb2d53f3ee78197ab
+
+
 ## [ForwardWidgets-huangxd] - 2025-08-17 03:18:06
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main